+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.52.0
 
 - Added `Wizaplace\SDK\Currency\CurrencyService::getAll`
+- Added `Wizaplace\SDK\Currency\CurrencyService::getByFilters`
+- Added `Wizaplace\SDK\Currency\CurrencyService::getByCountryCode`
 - Added `Wizaplace\SDK\Currency\CurrencyService::getCountries`
 - Added `Wizaplace\SDK\Currency\CurrencyService::addCountry`
 - Added `Wizaplace\SDK\Currency\CurrencyService::removeCountry`